test-requirements.txt
ansible-lint==5.4.0
molecule==4.0.1
molecule-docker==2.0.0
docker==5.0.3
yamllint==1.27.1
flake8==4.0.1
pytest==7.1.2
pytest-testinfra==6.8.0
tox==3.25.1
tox-gh-actions==2.9.1
netaddr==0.8.0
